requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0

ply==3.11

google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-python-client==2.187.0
google-api-core==2.28.1
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0

python-dotenv==1.0.0
requests==2.31.0

gunicorn==21.2.0
